Added image related loaders
