--- config/webpack/prod.js
@@ -1,9 +1,0 @@
-const config = require('./base');
-
-const prodConfig = {
-  entry: {
-    main: ['./app/main.js']
-  }
-};
-
-module.exports = Object.assign(config, prodConfig);
