+++ package.json
@@ -4,7 +4,7 @@
   "description": "",
   "main": "index.js",
   "scripts": {
-    "start": "webpack-dev-server --config config/webpack/dev.js",
+    "start": "webpack-dev-server --config config/webpack/development.js",
     "test": "echo \"Error: no test specified\" && exit 1",
     "stylelint": "stylelint 'app/**/*.css'"
   },
@@ -43,6 +43,8 @@
     "eslint-plugin-import": "^1.16.0",
     "eslint-plugin-jsx-a11y": "^2.2.3",
     "eslint-plugin-react": "^6.4.1",
+    "file-loader": "^0.9.0",
+    "image-webpack-loader": "^3.0.0",
     "mocha": "^3.1.2",
     "normalize.css": "^5.0.0",
     "postcss": "^5.2.5",
@@ -50,10 +52,12 @@
     "postcss-loader": "^1.0.0",
     "postcss-modules-values": "^1.2.2",
     "postcss-nested": "^1.0.0",
+    "raw-loader": "^0.5.1",
     "react-hot-loader": "^3.0.0-beta.6",
     "style-loader": "^0.13.1",
     "stylelint": "^7.5.0",
     "stylelint-config-standard": "^14.0.0",
+    "url-loader": "^0.5.7",
     "webpack": "^2.1.0-beta.25",
     "webpack-dev-server": "^2.1.0-beta.9"
   },
